Add default_parameters
