+++ papermill/inspection.py
@@ -9,8 +9,7 @@ import click
 
 from .iorw import get_pretty_path, open_notebook
 from .models import Parameter
-from .parameterize import find_first_tagged_cell_index
-from .translators import papermill_translators
+from .utils import any_tagged_cell
 
 
 def display_notebook_help(notebook_path):
@@ -28,23 +27,16 @@ def display_notebook_help(notebook_path):
     nb = open_notebook(notebook_path)
     pretty_path = get_pretty_path(notebook_path)
     click.echo("Usage: papermill [OPTIONS] {} [OUTPUT_PATH]".format(pretty_path))
-    click.echo("")
-    click.echo("  Parameters inferred for notebook '{}':".format(pretty_path))
+    click.echo("\n  Parameters inferred for notebook '{}':".format(pretty_path))
     
-    parameter_cell_idx = find_first_tagged_cell_index(nb, "parameters")
-    if parameter_cell_idx < 0:
+    if not any_tagged_cell(nb, "parameters"):
         click.echo("\n  No cell tagged 'parameters'")
         return
-    
-    parameter_cell = nb.cells[parameter_cell_idx]
-    kernel_name = nb.metadata.kernelspec.name
-    language = nb.metadata.kernelspec.language
 
-    translator = papermill_translators.find_translator(kernel_name, language)
-    params = translator.inspect(parameter_cell)
+    params = nb.metadata['papermill']['default_parameters']
     if params:
         for p in params:
             type_repr = p.inferred_type_name if p.inferred_type_name != "None" else "Unknown type"
-            click.echo("     {}: {} (default {})\t{}".format(p.name, type_repr, p.default, p.help))
+            click.echo("     {}: {} (default {})\t\t{}".format(p.name, type_repr, p.default, p.help))
     else:
-        click.echo("  Can't infer anything about this notebook's parameters")
+        click.echo("\n  Can't infer anything about this notebook's parameters. It may not have any parameter defined.")
